Added sanity check to make sure end date is AFTER start date!
Should fix problems pertaining to approving holiday for student editors...
The problems were caused when the system attempted to check vacation lapses
by examining the previous workplan and the current workplan. Afterward, it
would attempt to determine the vacation credit left for the user from the
start of the workplan to the end of their employment. Something that wouldn't
pass for users only employed a short while.